Additional out-of-the-box conditions demonstration
This commit adds the spring conventions that we apply on start.spring.io
to our simple generator. Such conditions require a few more dependencies
and, in particular, an InitializrMetadata instance that provides
information about available options.

The contributors that we added on the classpath need to know which
"platform version" is requested so the runner has been updated to
provide that information.

We can see that we have now two contributors that are competing with
each other so the next commit will replace our simple Maven build
contributor.